pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "figmaforge"
version = "1.0.0"
description = "Forge production-ready Angular components from Figma designs with AI-powered MCP integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "FigmaForge Team"}
]
keywords = ["figma", "angular", "component-generator", "mcp", "design-systems", "ai", "copilot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "mcp>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
figmaforge = "cli:cli"
figmaforge-sync = "cli:sync"
figmaforge-generate = "cli:generate"
figmaforge-init = "cli:init"
figmaforge-watch = "cli:watch"

[tool.setuptools]
packages = ["src"]
py-modules = ["cli", "mcp_stdio_server"]

